Move function selective_merge to util
