+++ lutris/installer/commands.py
@@ -13,22 +13,10 @@ from lutris import runtime
 from lutris.util import extract, disks, system
 from lutris.util.fileio import EvilConfigParser, MultiOrderedDict
 from lutris.util.log import logger
+from lutris.util import selective_merge
 
 from lutris.runners import wine, import_task
 from lutris.thread import LutrisThread
-
-
-def selective_merge(base_obj, delta_obj):
-    """ used by write_json """
-    if not isinstance(base_obj, dict):
-        return delta_obj
-    common_keys = set(base_obj).intersection(delta_obj)
-    new_keys = set(delta_obj).difference(common_keys)
-    for k in common_keys:
-        base_obj[k] = selective_merge(base_obj[k], delta_obj[k])
-    for k in new_keys:
-        base_obj[k] = delta_obj[k]
-    return base_obj
 
 
 class CommandsMixin(object):
