+++ lutris/util/__init__.py
@@ -1,1 +1,14 @@
 """ Misc common functions """
+
+
+def selective_merge(base_obj, delta_obj):
+    """ used by write_json """
+    if not isinstance(base_obj, dict):
+        return delta_obj
+    common_keys = set(base_obj).intersection(delta_obj)
+    new_keys = set(delta_obj).difference(common_keys)
+    for k in common_keys:
+        base_obj[k] = selective_merge(base_obj[k], delta_obj[k])
+    for k in new_keys:
+        base_obj[k] = delta_obj[k]
+    return base_obj
